% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_day.R
\name{water_day}
\alias{water_day}
\title{Get water day component of datetime(s)}
\usage{
water_day(x, start_month = 10)
}
\arguments{
\item{x}{vector of datetime objects}

\item{start_month}{integer specifying first month of water year (default = 10 for October)}
}
\value{
vector of water days as decimal numbers
}
\description{
A water day is the number of days past the start of a water year beginning with 1.
For example, if a water year is defined as starting in October, then the water day for
2007-09-30 is 365, 2007-10-01 is 1, 2007-10-02 is 2, etc.
}
\examples{
x <- seq.Date(from = as.Date("2007-09-30"), to = as.Date("2007-10-02"), by = "day"))
water_day(x)
}

