% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_year.R
\name{water_year}
\alias{water_year}
\title{Get water year component of datetime(s)}
\usage{
water_year(x, start_month = 10)
}
\arguments{
\item{x}{vector of datetime objects}

\item{start_month}{integer specifying first month of water year (default = 10 for October)}
}
\value{
vector of water years as decimal numbers
}
\description{
A water year is a 12-month period beginning in some month other than January.
The water year is designated by the calendar year in which it ends (e.g.
water year 2007 ends on 2007-09-30 if the starting month is October)
}
\examples{
x <- seq.Date(from = as.Date("2007-09-30"), to = as.Date("2007-10-02"), by = "day"))
water_year(x)
}

